+++ Source/Private/Get-NinjaOneKeyVaultInformation.ps1
@@ -1,0 +1,12 @@
+function Get-NinjaOneKeyVaultInformation {
+    <#
+        .SYNOPSIS
+            Retrieves NinjaOne connection and authentication information from the given Azure Key Vault.
+        .DESCRIPTION
+            Handles the reading of NinjaOne connection and authentication information to the given Azure Key Vault. This function is intended to be used internally by the module and should not be called directly.
+        .OUTPUTS
+            [System.Void]
+
+            Returns nothing.
+    #>
+}