+++ Source/Public/Get/Ticketing/Get-NinjaOneTicketBoards.ps1
@@ -25,10 +25,6 @@ function Get-NinjaOneTicketBoards {
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSReviewUnusedParameter', '', Justification = 'Uses dynamic parameter parsing.')]
     Param()
     begin {
-        if ($Script:NRAPIConnectionInformation.AuthMode -eq 'Client Credentials') {
-            throw ('This function is not available when using client_credentials authentication. If this is unexpected please report this to api@ninjarmm.com.')
-            exit 1
-        }
         $CommandName = $MyInvocation.InvocationName
         $Parameters = (Get-Command -Name $CommandName).Parameters
         $QSCollection = New-NinjaOneQuery -CommandName $CommandName -Parameters $Parameters